requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
asyncpg>=0.28.0
SQLAlchemy>=2.0.0
pydantic>=2.4.0
pytz>=2023.3
phonenumbers>=8.13.18
APScheduler>=3.10.0 